Fix filters (queries) constraining reports
This requires overwriting a file within the `group-by` dependency,
documented in the README. Hopefully, we will figure out how to push this
back into the `group-by` library. See
https://github.com/joemfb/group-by/pull/5

Also, I eliminated the search options, since they were only
'unfiltered', and that is not necessary for the `group-by` library,
which runs queries unfiltered in any case.