Remove query filter edit button
This feature didn't work, even in the original analytics dashboard, as
far as I can tell. You could bring up an editing interface, but they did
not get pushed back into the report. If we decide we want / can support
this feature, we could bring this back. Or something like it.