Add back settings screen, so can have a pie chart
This really should be baked into the design screen instead.

There is an option here for "parameters", but I cannot determine what
impact that has on the report. Asking Jianmin. The reason it is not
currently working is that, in the original analytics dashboard code,
Jianmin added `addParameter()` and `removeParameter` functions to the
malhar-dashboard dependency javascript.

Specifically, in the 'WidgetsSettingsCtrl':

```javascript
    $scope.addParameter = function () {
      var inserted = {
        name: '',
        value: ''
      };
      $scope.result.dataModelOptions.parameters.push(inserted);
    };

    // Delete parameter
    $scope.deleteParameter = function(index) {
      $scope.result.dataModelOptions.parameters.splice(index, 1);
    };

```