Create report uri before saving
This works around an apparent bug in slush-marklogic-spring-boot, which
did not return a response when creating a new document (so the browser
does not know the created uri). In general, it seemed wise not to rely
on the middle tier to pass through this information.

This is a very weak implementation, but should be more than enough for
our purposes.