Improve showing the entire generated query
Use the same code path to show it to the user as for sending it to the
server.

I started to refactor to build up a serializedQuery and use it directly everywhere
it is needed. Still need to refactor the filter queries to use it, and
then I can get rid of "generateQueryConfig" step