Turn saved queries back into filters
and fix other kinds of queries.

Still need to test with other kinds of queries, and also would like to
figure out a way to minimize this translation between 'filters' and
'queries'. It is made a bit more difficult because there is a different
data structure for queries vs. group-bys.