Isolate saved state from current working state
The app should track saved state separately. This will allow for
reversion. It also allows for the chart view to be tied to the saved
state, if desired (meaning that the chart will only update when a save
is made). Alternatively, the chart view could be tied to the working
state of the data.